requirements.txt
flask
flask-restful
flask-cors
flask-bcrypt
flask-jwt-extended
flask-sqlalchemy
flask-script
flask-migrate
werkzeug
gunicorn
